Catch unsupported build mode passed to build.py
When build.py is invoked with --build_only, it reuses the existing
controller source code to build the docker image. This assumes that the
necessary sieve instrumentation for building the controller in the
specified mode is already present in the source tree. But this limits
passing in 'ALL' mode to build.py to build all three controller image
variants at the same time. Adding in code changes to throw an error in
this case until this support is added.