Add a macos platform tag in node image
Running kind on macOS with node image (that has Sieve instrumentation)
built on a Linux machine is failing with errors - see
https://gist.github.com/jerrinsg/578fa112c232a72360367300be563901.
Building the node image locally on a Mac and using that image works
though. This commit therefore adds a new "macos" platform tag to node
image built for macOS so as to specifically use this image when running
on Mac.